+++ src/main/java/io/github/astrapi69/collection/list/ListExtensions.java
@@ -497,6 +497,38 @@ public final class ListExtensions
 	}
 
 	/**
+	 * Relocated the given element in the given {@link List} to the given index
+	 *
+	 * @param <T>
+	 *            the generic type of the elements
+	 * @param list
+	 *            the list that contains the element that have to be relocated
+	 * @param element
+	 *            the element to relocated
+	 * @param newIndex
+	 *            the new index to relocated
+	 */
+	public static <T> void relocate(List<T> list, T element, int newIndex)
+	{
+		Argument.notNull(element, "element");
+		Argument.notNull(list, "list");
+		int index = list.indexOf(element);
+		if (index < 0 || index == newIndex)
+		{
+			return;
+		}
+		list.remove(index);
+		if (list.size() <= newIndex)
+		{
+			list.add(element);
+		}
+		else
+		{
+			list.add(newIndex, element);
+		}
+	}
+
+	/**
 	 * Removes the first object from the given List.
 	 *
 	 * @param <T>
@@ -508,12 +540,7 @@ public final class ListExtensions
 	 */
 	public static <T> T removeFirst(final List<T> list)
 	{
-		Argument.notNull(list, "list");
-		if (!CollectionExtensions.isEmpty(list) && 0 < list.size())
-		{
-			return list.remove(0);
-		}
-		return null;
+		return removeFirstElement(list).orElse(null);
 	}
 
 	/**
@@ -549,11 +576,7 @@ public final class ListExtensions
 	 */
 	public static <T> T removeLast(final List<T> list)
 	{
-		if (!CollectionExtensions.isEmpty(list) && 0 < list.size())
-		{
-			return list.remove(list.size() - 1);
-		}
-		return null;
+		return removeLastElement(list).orElse(null);
 	}
 
 	/**